[DEV-5941] Move internal unit tests from `controllers` to `internal`
