+++ server/tests/internal/features/test_histogram.py
