+++ server/tests/internal/features/test_statistics.py
